feat(client): Option to trigger a new token gen
With this new option, users can create clients with a token in it!

Note that order matters when using this option since it will be
triggered with the current state of the `*client` struct, use it at the
end of a `New()` func.

An example:
```go
c, err := client.New("example",
	client.WithApiV2(),
	client.WithExpirationTime(1800),
	client.WithTokenFromKeys("KEY", "SECRET"), // this option has to be the last one
)
```

Signed-off-by: Salim Afiune Maya <afiune@lacework.net>