feat: introducing named profiles
```
As a Lacework CLI User with multiple Lacework Accounts,
I would like to be able to configure my Lacework CLI with multiple profiles,
So I can switch easily between accounts without having to reconfigure my configuration file.
```

This change is introducing named profiles and removing the parameter
`--config` since we will hardcode the config to be in
`$HOME/.lacework.toml`. Why? Since we can configure multiple profiles
today, we do not need to pass in any other config location.

Previous versions of the config are compatible to this change so there
are no breaking changes.

All environment variables previously defined continues to work in
addition to the new environment variable `LW_PROFILE` that switches the
profile to use.

Closes https://github.com/lacework/go-sdk/issues/36

Signed-off-by: Salim Afiune Maya <afiune@lacework.net>