refactor(api): request and response log messages
To help the Go api client to log information messages that do not
overload the output we display, we are suppressing headers and body of
requests on `INFO` log level. If users wants to see those details, they
will have to set `DEBUG` to sniff inside the request and response
headers and bodies.

Additionally, we are adding a number of log messages, both INFO and
DEBUG, so that we can present more logging information.

Signed-off-by: Salim Afiune Maya <afiune@lacework.net>