feat(cli): add --details flag to vulnerability cmd
This change is adding a new flag to the vulnerability command called
`--details`, this flag will extend the report shown to the end-user.

The new information presented with this flag is the list of CVE's with
its severity, version and, if any, fix version, plus the layer that the
vulnerability was introduced by.

Example:
```
$ lacework vulnerability report sha256:123ab...xyz --details
```

This flag is available to all sub-commands inside the `vulnerability`
cmd. (`scan run`, `scan show`, and `report` as shown above)

Closes https://github.com/lacework/go-sdk/issues/82

Signed-off-by: Salim Afiune Maya <afiune@lacework.net>