feat(cli): create integration sub-command
We are adding the ability to create external integration via the CLI, as
of this commit, we can create Docker Hub and AWS Config integrations.
The code has comments of the rest of the integrations available in the
platform, such integration will be added on further changes.

Example: Create a Docker Hub integration
```
$ lacework integration create
? Choose an integration type to create:  Docker Hub
▸ Name:  techallylw docker hub from CLI
▸ Username:  techallylw
▸ Password:  ***********************
▸ Limit by Tag:  *
▸ Limit by Label:  *
▸ Limit by Repository:
▸ Limit Number of Images per Repo:  5
The integration was created.
```

Signed-off-by: Salim Afiune Maya <afiune@lacework.net>