feat(cli): preconfigure using key JSON file (WebUI)
This change is adding an extra flag called `--json_file`, short that `-j`
that will allow users load a generated API key JSON file from the WebUI.

Example: On this example, we create a set of API keys by logging in into
the Lacework WebUI, then download the keys and pass them to the Lacework
CLI to preconfigure it.
```
$ lacework configure --json_file /path/to/EXAMPLE_D3EFA214SD12DFSGRD3EFA214SD12DFSGR.json
```

The result of the above command is to only provide the "Account"
information, the API access key and secret will be preloaded.

Signed-off-by: Salim Afiune Maya <afiune@lacework.net>