feat: understand vuln reports with 0 vulnerabilities
Make the Go API and the CLI to understand new payloads from the vuln
reports API endpoints that displays reports with 0 vulnerabilities,
we are now displaying a message to the user that explains that the
report has no vulnerabilities

The output shown to the end-user is for reports with 0 vulnerabilities:
```
$ lacework vul report sha256:32c5a42735581453d1e9ea664604fa81f4b635459dc88799ee1f8c373dca4819
Great news! This container image has no vulnerabilities.
```

And for the JSON format:
```
$ lacework vul report sha256:32c5a42735581453d1e9ea664604fa81f4b635459dc88799ee1f8c373dca4819 --json
{
  "critical_vulnerabilities": 0,
  "fixable_vulnerabilities": 0,
  "high_vulnerabilities": 0,
  "info_vulnerabilities": 0,
  "low_vulnerabilities": 0,
  "medium_vulnerabilities": 0,
  "scan_status": "Success",
  "total_vulnerabilities": 0
}
```

Closes https://github.com/lacework/go-sdk/issues/123

Signed-off-by: Salim Afiune Maya <afiune@lacework.net>