feat(ux): generate package-manifest command
This Pull Request is improving the way we scan package-manifests, this is all
about User Experience (UX).

We want to help our users to adopt our CI/CD story and make it easy for them
to scan their package manifests, with that mentality we are introducing a couple
new feature to the Lacework CLI.

* Generation of Package Manifest
Generating a package manifest could be confusing for some users, if you are an
expert you will probably have your own Python scripts ready to be piped to our
APIs, but if you are not, we have a new command for you:
```
root@7c42aed31953:/# lacework vulnerability host generate-pkg-manifest
{
  "os_pkg_info_list": [
    {
      "os": "ubuntu",
      "os_ver": "18.04",
      "pkg": "adduser",
      "pkg_ver": "3.116ubuntu1"
    },
    {
      "os": "ubuntu",
      "os_ver": "18.04",
      "pkg": "zlib1g",
      "pkg_ver": "1:1.2.11.dfsg-0ubuntu2"
    }
  ]
}
```
This command will automatically detect any supported package manager and generate
the package manifest for you!

![tenor-111571833](https://user-images.githubusercontent.com/5712253/93834021-af241600-fc37-11ea-9a76-71f168ddd79c.gif)

* Scan Local Host
Let us take it one step further! Since we now have a way to generate a package manifest
for you, we have also added a new flag `--local` that you can use to automatically generate
the package manifest on a running host and submit a scan, all at once!

```
root@7c42aed31953:/# lacework vulnerability host scan-pkg-manifest --local
        CVE        |  SEVERITY  | SCORE |   PACKAGE   |       VERSION       |    FIX VERSION
-------------------+------------+-------+-------------+---------------------+---------------------
  CVE-2016-6313    | High       |   5.3 | libgcrypt20 | 1.8.1-4ubuntu1.2    | 0:1.7.2-2ubuntu1
  CVE-2019-13627   | Medium     |   6.3 | libgcrypt20 | 1.8.1-4ubuntu1.2    | 0:1.8.1-4ubuntu1.2
  CVE-2018-0495    | Low        |   4.7 | libgcrypt20 | 1.8.1-4ubuntu1.2    | 0:1.8.1-4ubuntu1.1
  CVE-2016-2781    | Low        |   6.5 | coreutils   | 8.28-1ubuntu1       |
  CVE-2019-12904   | Low        |   5.9 | libgcrypt20 | 1.8.1-4ubuntu1.2    |
  CVE-2019-9923    | Low        |   7.5 | tar         | 1.29b-2ubuntu0.1    |
  CVE-2018-20482   | Low        |   4.7 | tar         | 1.29b-2ubuntu0.1    |
  CVE-2019-18276   | Low        |   7.8 | bash        | 4.4.18-2ubuntu1.2   |
  CVE-2017-8283    | Negligible |   9.8 | dpkg        | 1.19.0.5ubuntu2.3   |
  CVE-2018-1000654 | Negligible |   5.5 | libtasn1-6  | 4.13-2              |
  CVE-2016-9401    | Negligible |   5.5 | bash        | 4.4.18-2ubuntu1.2   | 0:4.4-5ubuntu1
  CVE-2018-7738    | Negligible |   7.8 | util-linux  | 2.31.1-0.4ubuntu3.6 |
```

![tenor-263751943](https://user-images.githubusercontent.com/5712253/93834123-f7dbcf00-fc37-11ea-9c16-3dcafec14f05.gif)

Signed-off-by: Salim Afiune Maya <afiune@lacework.net>