fix: false positive results in pkg manifest scan
The API response coming from the Lacework server contains too much
information that could confuse our users, this change is cleaning the
response by removing all vulnerabilities that don't match, that is,
those vulnerabilities that have the evaluation status equal to
`VULNERABLE`.

This fix is modifying both, the human-readable and JSON outputs.

Signed-off-by: Salim Afiune Maya <afiune@lacework.net>