refactor(cli): abstract rendering tables (human-readable) (#263)
This refactor is adding a new function named `renderSimpleTable()` used
to render any simple table within the Lacework CLI, every command should
leverage this function unless there are extra customizations required,
if so, use instead `renderCustomTable()`. The benefit of this function
is the ability to switch/update the look and feel of the human-readable
format across the entire project.

The function `renderSimpleTable()` receives two arguments, the headers
(`[]string`), and the rows. (`[][]string`)

For example, to render a simple table like the following one:
```
  KEY    VALUE
-------+---------
  key1   value1
  key2   value2
  key3   value3
```

The code would look like:
```go
cli.OutputHuman(
	renderSimpleTable(
		[]string{"KEY", "VALUE"},
		[][]string{
			[]string{"key1", "value1"},
			[]string{"key2", "value2"},
			[]string{"key3", "value3"},
		}),
)
```

Signed-off-by: Salim Afiune Maya <afiune@lacework.net>