feat(cli): enable agent install command
We are ready to enable this new command to help onboarding new Lacework
agents! Users can now install agents with:

```
$ lacework agent install
```

When this command is executed without any additional flag, an interactive prompt will be
launched to help gather the necessary authentication information to access the remote host.

To authenticate to the remote host with a username and password.
```
$ lacework agent install <host> --ssh_username <your-user> --ssh_password <secret>
```

To authenticate to the remote host with an identity file instead.

```
$ lacework agent install <user@host> -i /path/to/your/key
```

To provide an agent access token of your choice, use the command 'lacework agent token list',
select a token and pass it to the '--token' flag.

```
$ lacework agent install <user@host> -i /path/to/your/key --token <token>
```

Signed-off-by: Salim Afiune Maya <afiune@lacework.net>