feat(cli): adds telemetry
As a Lacework engineer,
I need to have observability of how our customers are using the Lacework CLI,
where are they having issues, what are the most used commands,
and which features are they using the most,
So I can understand and make decisions about priorities of bug fixes,
new features, and improvements.

This change is implementing sending tracing data to Honeycomb.

**Acceptance Criteria**

After completing this story, we should be able to understand and answer the following questions:
* List of most used commands
* List of commands with the most number of errors
* List of most common errors
* Usage of CLI features (polling mechanism, daily version check, etc.)
* List of active/installed CLI versions
* List of most active customers
* List of most active API key/User per customer
* List of most used profiles

Signed-off-by: Salim Afiune Maya <afiune@lacework.net>