feat(cli): gen-pkg-manifest detect running kernel
The default behavior of most linux distros is to keep the last N kernel
packages installed for users that need to fallback in case the new
kernel do not boot.  However, the presence of the package does not mean
that kernel is active.  We must continue to allow the standard kernel
package preservation behavior without providing false-positives of
vulnerabilities that are not active.

This feature detects the active kernel and removes any other
installed-and-inactive kernel from the generated package manifest.

Signed-off-by: Salim Afiune Maya <afiune@lacework.net>