fix: implement both ECR auth methods
A new change was released last week that changed the SCHEMA of an ECR
Integration, a new required field named `AWS_AUTH_TYPE` was added that
broke our integrations, this is the fix for such bug.

Additionally, we have implemented both authentication methods, IAM Role
and Access Key.

JIRA: ALLY-371

Signed-off-by: Salim Afiune Maya <afiune@lacework.net>