feat(cli): Automate migration of v2 configuration
**User Story**
As a Lacework CLI User, I won't like to be disrupted during the
migration of the Lacework CLI from APIv1 to APIv2, So I can
continue accessing my Lacework data at all times and I don't have
to reconfigure all my profiles.

**Feature Description**
During the migration of our tools and integrations (in this case,
the Lacework CLI) we won't disrupt any of our customers, instead,
we are codifying and automating the migration by steps, tiny steps
that won't be noticed by our customers.

The first step of our plan is to automate the migration of the
configuration of existing customers so that the Lacework CLI can access
both, APIv1 and APIv2 at the same time for a short period of time (while
we slowly migrate existing functionality to APIv2). To do that, we send
an APIv1 request that describes the type of Lacework Account that our
customers have, there are two types, standalone accounts, and
organizational accounts.

For standalone accounts, the new configuration version 2 might look
practically the same as version 1, but the fact that we add the version
to the configuration will indicate that we no longer need to migrate it.

Example of version 1 standalone config:
```toml
[default]
  account = "standalone"
  api_key = "STANDALONE_13C8910D0BDCDBD7AFD4355A1C5284104AAA2AE5953938C"
  api_secret = "_123abc32a752e5f94651bcb6e6c798f1"
```

Example of version 2 standalone config:
```toml
[default]
  account = "standalone"
  api_key = "STANDALONE_13C8910D0BDCDBD7AFD4355A1C5284104AAA2AE5953938C"
  api_secret = "_123abc32a752e5f94651bcb6e6c798f1"
  version = 2
```

For organizational accounts, the new configuration version 2 will have a
new field named `subaccount`, and the field account will change to point
to the primary account (unique account URL).

Example of version 1 organizational account config:
```toml
[default]
  account = "sub-account"
  api_key = "SUBACCOUNT_13C8910D0BDCDBD7AFD4355A1C5284104AAA2AE5953938C"
  api_secret = "_123abc32a752e5f94651bcb6e6c798f1"
```

Example of version 2 standalone config:
```toml
[default]
  account = "org-account"
  subaccount = "sub-account"
  api_key = "SUBACCOUNT_13C8910D0BDCDBD7AFD4355A1C5284104AAA2AE5953938C"
  api_secret = "_123abc32a752e5f94651bcb6e6c798f1"
  version = 2
```

For more details, refer to the migration document of the EPIC.

**Acceptance Criteria**
Users using this feature will start the process of migrating their
Lacework CLI configuration. If we do this right, our customers won't
notice that this change happened. Additionally, we will be able to start
leveraging APIv2 endpoints from this new version of the CLI.

This ticket adds a new Honeycomb Feature Event (metric) to detect
when customers have been migrated to the new CLI configuration v2.

Feature Name: `migrate_config_v2`

JIRA https://lacework.atlassian.net/browse/ALLY-376

Signed-off-by: Salim Afiune Maya <afiune@lacework.net>