feat(cli): cache host vuln assessments for 1 hour
This is the first feature we want implement where we cache an asset, in
this case, a host vulnerability assessment for one hour.

One main benefit is the ability to run further commands with additional
flags to manipulate the output of the assessment without having to fetch
the asset from the APIs again.

Why one hour!? Because our agent runs a scan and returns results (the
assessment) every day (24 hours)

Signed-off-by: Salim Afiune Maya <afiune@lacework.net>