feat(cli): cache compliance reports for 30 minutes
As a Lacework CLI user,
I want to inspect compliance reports via the CLI multiple times
without getting to API rate limits,
So I can create workflows internally and navigate my reports with
additional flags and filters

As a Lacework engineer,
want tools and integrations to cache data as much as possible without
compromising consistency,
So other services, such as the API server, don't get overwhelmed with
the same API requests that are immutable

When a user is trying to visualize a compliance report via the CLI, we
shouldn’t ask the APIs for the same report on every command execution,
especially not since the report won’t change between two commands run
within a few minutes. We should be able to use the Caching mechanism
inside the Lacework CLI to cache compliance reports and let users
navigate them faster and reduce the load to our API server.

We have already done this for host vulnerability assessments.

Signed-off-by: Salim Afiune Maya <afiune@lacework.net>