feat(cli): be able to change global LW_LQL_OPERATOR
The `eq` operator allows you to specify a value that the field values
of the result must be equal to. The `ne` operator means not equal to.

The `in` operator allows you to specify multiple values in the values
field of the filters. The field values of the result must match one of
the values. The `not_in` operator is the opposite of `in`.

The `like` operator allows you to specify a pattern that the field
values of the result must match. The `not_like` operator is the
opposite of `like`.

The `ilike` operator works similar to like but it makes the match case
insensitive. The `not_ilike` operator is the opposite of `ilike`.

(default) The `rlike` operator matches the specified pattern represented
by regular expressions (more info on `RLIKE` see https://docs.snowflake.com/en/sql-reference/functions/rlike.html).
The `not_rlike` operator is the opposite of `rlike`.

The `gt` operator allows you to specify a value that the field values of
the result must be greater than. The `lt` (less-than) operator is the
opposite of `gt`.

The `ge` operator allows you to specify a value that the field values of
the result must be greater than or equal to. The `le` (less-than-or-equal-to)
operator is the opposite of `ge`.

Signed-off-by: Salim Afiune Maya <afiune@lacework.net>