feat(cli): update query by loading it from the platform
As a Lacework CLI user,
I want to update a query without having to type the query again,
So I can modify my queries faster via inline modifications.

This change is adding the ability to load a query into the default
editor from the user so that they can easily make inline modifications.

Simply run the command `lacework query update [query_id]` and an editor
will open with the query preloaded, modify it, save and close the file
for the Lacework CLI to apply those modifications.

Signed-off-by: Salim Afiune Maya <afiune@lacework.net>