add newResponse for function SearchAllPages
This is necessary as using the current response to iterate over pages
corrupts data that are slices, such as the tags within imageInfo. This is beacause
the memory address from the previous response will still contain data from the previous request.
By creating a new response using the nextPage url we get a new spot in memory to put the next
requests data.

Signed-off-by: aircraft-cerier <58537442+aircraft-cerier@users.noreply.github.com>