feat(cdk): pass env variable LW_COMPONENT_NAME
To integrate the Soluble CLI into the Lacework CLI we are planning to add it as a
component (an extension) so that customers can discover and use it out-of-the-box.

This new environmental variable will allow us to inject the name of the component
at runtime so that all the help messages reflect the correct component name.

Depends on https://github.com/soluble-ai/soluble-cli/pull/431

Signed-off-by: Salim Afiune Maya <afiune@lacework.net>