fix(install.ps1): modify User $Path correctly
When an non-admin Windows user would install the Lacework CLI, we were
modifying the User's $Path environment variable incorrectly. In fact, we
were overriding the entire $Path to be the System's $Path (at the Machine)

This change is fixing this issue by properly updating the Users' or
System wide $Path environment variable.

Jira:
- https://lacework.atlassian.net/browse/LINK-812
- https://lacework.atlassian.net/browse/ALLY-1103

Signed-off-by: Salim Afiune Maya <afiune@lacework.net>