feat(cli): parallelize installation with goroutines (#948)
Currently, both `lacework agent aws-install ec2ic` and
`lacework agent aws-install ec2ssh` capture instances
in series. This is fine for a low number of instances that
are to be captured, but at the scale of 10k instances, this
quickly breaks down because it takes too long. 

This commit uses goroutines to parallelize installation.
And since the AWS API has rate limits, we want to limit
the number of concurrent threads making requests.
This commit also adds a new flag, `--max_parallelism`,
to set the number of workers. The default value of this
flag is 50. 

This commit uses github.com/gammazero/workerpool
for worker pool management.

Fixes RAIN-38534
Fixes RAIN-38695

Signed-off-by: nschmeller <nick.schmeller@lacework.net>