fix(cdk): parse global flags when running components
Before this change, when someone would try to use global flags such as
`--profile`, `--debug`, etc. while running a component, the CLI would
bypass the parsing of all flags and it would send all arguments directly
to the underlying component. This caused problems when switching
profiles, activating debug logs, etc. (no global flag was working)

This change is implementing a parsing mechanism that will filter all
real global flags from the provided arguments and pass the rest as
arguments to the component. The filtered global flags are also parsed
and applied.

Signed-off-by: Salim Afiune Maya <afiune@lacework.net>