feat(cli): find and fix GCP instances w/o agents (#1022)
This commit is the second milestone for Project Capture The
Flag. The first milestone was in commit
4e507c0b89edde2a8f04e03581a92aae9789bdc9.

This commit adds a new subcommand to the Lacework CLI
called `gcp-install`. The user should invoke this command as
`lacework agent gcp-install`. `gcp-install` is currently a
no-op and is used as a directory for further subcommands.

This commit adds one further subcommands to `gcp-install`,
called `osl`. The user should invoke it as
`lacework agent gcp-install osl <gcp_username>`.

The `osl` subcommand uses GCE OSLogin to send SSH keys
to the user's GCP account. The user can then SSH into all
GCE instances which have `oslogin` enabled in their metadata.
The user must have the requisite permissions to use `OSLogin`,
https://pomba.net/2022/05/gcp-ssh-into-vms-as-service-account-when-oslogin-is-enabled/
is a useful guide for setting up the IAM permissions required to
use service accounts for this purpose.

`osl` has the limitation that it only works for instances with
public IP addresses that are open to the internet on port 22. 

Fixes RAIN-37376
Fixes RAIN-43512

Signed-off-by: nschmeller <nick.schmeller@lacework.net>