fix(cli): set correct SSH user for Amazon Linux 2
This commit fixes an error in the CLI where the SSH user was set
incorrectly for Amazon Linux 2. This commit also adds a new unit test to
verify this case. This commit also adds debug logs for the AMI for each
instance captured to give more visibility for these types of problems in
the future.

Fixes RAIN-43934

Signed-off-by: nschmeller <nick.schmeller@lacework.net>