Sleep at end of for loop
Discussion from GitHub PR

Here are the tradeoffs as I see them:

Pros:
Only sleep on failure (if customer has instance profile already
associated, this saves them a minute of execution)

Cons:
The last minute of sleep isn't useful, since we've already tried the
command and failed, so we're waiting a minute just to return an error.

I think the solution that has the best of both worlds is to wrap the
sleep with a conditional that executes the sleep if it's not the last
iteration in the loop. Which seems a little messy to me, but sleeps
aren't exactly pristine code in the first place.

Signed-off-by: nschmeller <nick.schmeller@lacework.net>