feat(cli): CTF `aws-install ec2ssm` (#1045)
This commit is the third milestone for Project Capture The
Flag. The second milestone was in commit
08edf28a69693926b5b01be54750ca61b671942a.

This commit adds a new subcommand to the `aws-install`
command in the Lacework CLI, called `ec2ssm`. The
user should invoke this command as
`lacework agent aws-install ec2ssm`.

The `ec2ssm` subcommand uses AWS SSM to install the
agent on target instances in the user's AWS account.
If the target instances are not setup to be managed with
SSM, then `ec2ssm` will create an AWS IAM role and
instance profile with the required policies attached, and
associate it with the target instances. `ec2ssm` will then
teardown these resources before exiting. The user should
have IAM full access permissions for the EC2, IAM, and SSM
services.

`ec2ssm` has the limitation that target instances must have
the SSM agent running. This will be true for most AWS
supported AMIs, as well as many others.

Fixes RAIN-38281

Signed-off-by: nschmeller <nick.schmeller@lacework.net>