refactor(cli): relative and natural time range for vuln list cmd
To customize the time range use use '--start', '--end', or '--range'.

The start and end times can be specified in one of the following formats:

    A. A relative time specifier
    B. RFC3339 date and time
    C. Epoch time in milliseconds

Or use a natural time range like.

    lacework vuln container list --range yesterday

The natural time range of 'yesterday' would represent a relative start time of '-1d@d'
and a relative end time of '@d'.

Signed-off-by: Salim Afiune Maya <afiune@lacework.net>