feat(cli): improve caching using hash of filters
When caching, we depend on a key to store and retrieve data, this cache
key should change depending on filters that will affect the data
returned by our APIs.

This change implements a function `hash()` to generate unique hashes of
arbitrary values in Go which is useful to generate unique cache keys of
filters applied to commands.

It also implements a second function `generateContainerVulnListCacheKey()`
that returns the cache key for vulnerability assessments. The criteria to
generate this cache key is to use the prefix
`vulnerability/container/v2_{HASH}` with a hash value at the end of the
prefix. The hash is generated based of filters that, if changed by the
user, will change the data returned from our APIs.

Signed-off-by: Salim Afiune Maya <afiune@lacework.net>