feat(cli): `ec2ssm` no longer reuses IAM infra
`ec2ssm` currently will reuse existing IAM infra from a previous
invocation. This was useful in avoiding extra IAM operations in the
event of an interrupt or a crash. However, this functionality meant that
`ec2ssm` was not idempotent and could not be run concurrently. This
property is true now.

This commit:

* adds a random identifier to the names of new IAM roles and profiles
* makes an AWS API call to disassociate instance profiles on exit
* removes checks for existing IAM roles and profiles

Fixes RAIN-47837

Signed-off-by: nschmeller <nick.schmeller@lacework.net>