chore(cli): add info about false-positive virus on Windows
When we upgraded to go version `1.21`, Chocolatey started flagging our binaries:

https://community.chocolatey.org/packages/lacework-cli#virus

![Screenshot 2023-09-08 at 11 24 25 AM](https://github.com/lacework/go-sdk/assets/5712253/7a61981b-dbae-43c6-8510-52eb44a5893a)

This is a common false-positive so this PR is adding this "banner" to explain:

---

**Important Notice**

Since the release of the Lacework CLI version `1.30.0` we noticed that the virus scan reports positive findings. We
have analyzed the generated binaries and we have not found any virus.

This is a **common false-positive** in the Golang community for Windows binaries. For details, go to https://golang.org/doc/faq#virus

---

Signed-off-by: Salim Afiune Maya <afiune@lacework.net>