fix(api): avoid round numbers in responses
By using the `UseNumber()` function of the decoder https://pkg.go.dev/encoding/json#Decoder.UseNumber
we are causing the Decoder to unmarshal a number into an interface{} as a
Number instead of as a float64.

This fixes the problem when the CLI returns the raw response of long
numbers like the `mid`.

Added tests that failed with:
```
=== FAIL: api TestDoDecoderLooongNumbers (0.00s)
    http_test.go:61:
        	Error Trace:	/Users/afiune/github/go-sdk/api/http_test.go:61
        	Error:      	Not equal:
        	            	expected: "5016581821971911018"
        	            	actual  : "5.016581821971911e+18"

        	            	Diff:
        	            	--- Expected
        	            	+++ Actual
        	            	@@ -1 +1 @@
        	            	-5016581821971911018
        	            	+5.016581821971911e+18
        	Test:       	TestDoDecoderLooongNumbers
```

Now passing.

https://lacework.atlassian.net/browse/GROW-2459

Signed-off-by: Salim Afiune Maya <afiune@lacework.net>