chore(goimports): change imports-check from go list to find
The previous imports-check target uses `go list` to find code to scan.
The problem is that `go list` produces a listing of directories to input
into goimports.

/a/b/c
/a/b/c/d
/a/b/c/e ---> example we want to exclude

In the example listing above, we want to exclude `e`. That was previously
done by excluding `e` via a `grep -v` command that followed `go list`.
The result is that the path with `e` on the end is excluded, but `/a/b/c`
is still scanned causing the excluded path to get scanned anyway. This
commit swaps all this out for a find command that excludes the paths and
returns filenames as input, not directories.