feat(cdk): expose cache to components (#1441)
* feat(GROW-2498): expose cache to components

* chore(goimports): change imports-check from go list to find

The previous imports-check target uses `go list` to find code to scan.
The problem is that `go list` produces a listing of directories to input
into goimports.

/a/b/c
/a/b/c/d
/a/b/c/e ---> example we want to exclude

In the example listing above, we want to exclude `e`. That was previously
done by excluding `e` via a `grep -v` command that followed `go list`.
The result is that the path with `e` on the end is excluded, but `/a/b/c`
is still scanned causing the excluded path to get scanned anyway. This
commit swaps all this out for a find command that excludes the paths and
returns filenames as input, not directories.

* chore: add ci build of go-component test resources

* chore: add integration test binaries back

These can be refreshed using the make cdk-go-component-ci command,
however it'll cause local integration tests to fail unless folks rebuild
these manually everytime on their machines. In addition, its slowing the
integration tests down substantially on an already long-running job.
These resources should be fairly static so we're making the choice to
include these resources statically.

* refactor: update cdk cache interface based on feedback